pdc: move metadata handler to dedicated file. save ipfs on request entity
