+++ polygon-digital-carbon/schema.graphql
@@ -416,6 +416,7 @@ type C3RetireRequest @entity {
   retire: Retire
   provenance: ProvenanceRecord
   tokenURI: String!
+  retirementMetadata: C3RetirementMetadata
 }
 
 # C3Project and C3RetirementMetadata are used for file datasources for C3 metadata
