+++ polygon-digital-carbon/src/RetirementHandler.ts
@@ -15,9 +15,9 @@ import { incrementAccountRetirements, loadOrCreateAccount } from './utils/Accoun
 import { loadCarbonCredit, loadOrCreateCarbonCredit } from './utils/CarbonCredit'
 import { loadOrCreateCarbonProject } from './utils/CarbonProject'
 import { loadRetire, saveRetire } from './utils/Retire'
-import { log, Bytes, json, dataSource } from '@graphprotocol/graph-ts'
+import { log } from '@graphprotocol/graph-ts'
 import { loadOrCreateC3RetireRequest, loadC3RetireRequest } from './utils/C3'
-import { C3RetirementMetadata, C3MetadataProject, Token, TokenURISafeguard } from '../generated/schema'
+import { Token, TokenURISafeguard } from '../generated/schema'
 import { getC3RetireRequestId } from '../utils/getRetirementsContractAddress'
 import { BridgeStatus } from '../utils/enums'
 import { loadOrCreateToucanBridgeRequest } from './utils/Toucan'
@@ -340,7 +340,7 @@ export function completeC3RetireRequest(event: EndAsyncToken): void {
           log.error('Retrieved tokenURI is null or empty for nft index {}', [event.params.nftIndex.toString()])
         } else {
           request.tokenURI = tokenURI
-
+          request.retirementMetadata = tokenURI
           C3RetirementMetadataTemplate.create(tokenURI)
         }
       }
@@ -376,71 +376,10 @@ export function handleVCUOMetaDataUpdated(event: VCUOMetaDataUpdated): void {
       const tokenURI = event.params.url
 
       request.tokenURI = tokenURI
+      request.retirementMetadata = tokenURI
       request.save()
 
       C3RetirementMetadataTemplate.create(tokenURI)
     }
-  }
-}
-
-export function handleC3RetirementMetadata(content: Bytes): void {
-  let c3RetirementMetadata = new C3RetirementMetadata(dataSource.stringParam())
-  const value = json.fromBytes(content).toObject()
-  if (value) {
-    const transferee = value.get('transferee')
-    const reason = value.get('reason')
-    const projectId = value.get('projectId')
-    const projectAddress = value.get('projectAddress')
-    const amount = value.get('amount')
-    const vintage = value.get('vintage')
-    const owner = value.get('owner')
-    const uncheckedProject = value.get('project')
-    const image = value.get('image')
-    const pdf = value.get('pdf')
-    const nftRegistry = value.get('nftRegistry')
-
-    if (transferee) c3RetirementMetadata.transferee = transferee.toString()
-    if (reason) c3RetirementMetadata.reason = reason.toString()
-    if (projectId) c3RetirementMetadata.projectId = projectId.toString()
-    if (projectAddress) c3RetirementMetadata.projectAddress = Bytes.fromHexString(projectAddress.toString())
-    if (amount) c3RetirementMetadata.amount = amount.toBigInt().toI32()
-    if (vintage) c3RetirementMetadata.vintage = vintage.toString()
-    if (owner) c3RetirementMetadata.owner = Bytes.fromHexString(owner.toString())
-    if (image) c3RetirementMetadata.image = image.toString()
-    if (pdf) c3RetirementMetadata.pdf = pdf.toString()
-    if (nftRegistry) c3RetirementMetadata.nftRegistry = nftRegistry.toString()
-
-    if (uncheckedProject) {
-      const project = uncheckedProject.toObject()
-      let projectEntity = new C3MetadataProject(c3RetirementMetadata.id)
-      const projectName = project.get('name')
-      const project_id = project.get('project_id')
-      const project_type = project.get('project_type')
-      const registry = project.get('registry')
-      const region = project.get('region')
-      const country = project.get('country')
-      const methodology = project.get('methodology')
-      const period_start = project.get('period_start')
-      const period_end = project.get('period_end')
-      const ac = project.get('ac')
-      const uri = project.get('uri')
-
-      if (projectName) projectEntity.name = projectName.toString()
-      if (project_id) projectEntity.project_id = project_id.toString()
-      if (project_type) projectEntity.project_type = project_type.toString()
-      if (registry) projectEntity.registry = registry.toString()
-      if (region) projectEntity.region = region.toString()
-      if (country) projectEntity.country = country.toString()
-      if (methodology) projectEntity.methodology = methodology.toString()
-      if (period_start) projectEntity.period_start = period_start.toString()
-      if (period_end) projectEntity.period_end = period_end.toString()
-      if (ac) projectEntity.ac = ac.toBool()
-      if (uri) projectEntity.uri = uri.toString()
-
-      projectEntity.save()
-      c3RetirementMetadata.project = projectEntity.id
-    }
-
-    c3RetirementMetadata.save()
   }
 }
