+++ polygon-digital-carbon/subgraph.template.yaml
@@ -626,7 +626,7 @@ templates:
     mapping:
       apiVersion: 0.0.7
       language: wasm/assemblyscript
-      file: ./src/RetirementHandler.ts
+      file: ./src/MetadataHandler.ts
       handler: handleC3RetirementMetadata
       entities:
         - C3RetireRequest
